Updated to LWJGL 3, including input and font rewrites

+++ src/uk/co/newagedev/hieranarchy/ui/ScrollPane.java
@@ -39,7 +39,7 @@ public class ScrollPane extends Component {
 				bar.update();
 			}
 		}
-		int mx = Mouse.getMouseX(), my = Mouse.getMouseY();
+		int mx = (int) Mouse.getMouseX(), my = (int) Mouse.getMouseY();
 		boolean hovering = (scrollBars[0] != null ? (my > getDisplayLocation().getY() + getHeight() && my < getDisplayLocation().getY() + getHeight() + 15) || scrollBars[0].isHeld() : false) || (scrollBars[1] != null ? (mx > getDisplayLocation().getX() + getWidth() && mx < getDisplayLocation().getX() + getWidth() + 15) || scrollBars[1].isHeld() : false);
 		if ((new Rectangle((int) getDisplayLocation().getX(), (int) getDisplayLocation().getY(), getWidth(), getHeight())).contains(Mouse.getMouseX(), Mouse.getMouseY())) {
 			pane.update();
